Set resource path to be relative index.html
Per the create-react-app documentation:
https://github.com/facebookincubator/create-react-app/blob/master/packages/react-scripts/template/README.md#building-for-relative-paths

Generally, we won't need this since we're going to be publishing it at
the root, but it's handy if we want to push it anywhere else for testing
or debugging purposes, we don't have to worry about any weird pathing
issues.

If we decide to add more content (like, a blog or something) and use
client side routing, we will have to revisit this though.